+++ dags/sql/create_order_table.sql
@@ -4,5 +4,5 @@ create table ORDERS (
     ord_id varchar PRIMARY KEY,
     cust_id varchar,
     prod_id varchar,
-    count numeric
+    count smallint
 );
